requirements.txt
asgiref==3.5.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.2
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
filelock==3.4.2
google-api-core==2.7.1
google-auth==2.6.2
google-cloud-vision==2.7.2
googleapis-common-protos==1.56.0
grpcio==1.44.0
grpcio-status==1.44.0
gunicorn==20.0.4
idna==3.3
Pillow==9.0.1
pipenv==2022.1.8
platformdirs==2.5.0
proto-plus==1.20.3
protobuf==3.19.4
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2021.3
requests==2.27.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.9
virtualenv==20.13.1
virtualenv-clone==0.5.7
whitenoise==5.2.0
